splitting up client.py and for now disabling isort.
