+++ .pre-commit-config.yaml
@@ -14,12 +14,6 @@ repos:
         language: system
         types: [python]
         require_serial: true
-      - id: isort
-        name: isort
-        entry: poetry run isort
-        language: system
-        types: [python]
-        require_serial: true
       - id: flake8
         name: flake8
         entry: poetry run flake8
